pyproject.toml
[tool.poetry]
name = "meta-yovisto"
version = "0.1.0"
description = ""
authors = ["Robert Meissner <robert.meissner@codecentric.de>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
fastapi = "^0.78.0"
starlette-context = "^0.3.3"
python-dotenv = "^0.20.0"
rdflib = "^6.1.1"
nltk = "^3.7"
treelib = "^1.6.1"
uvicorn = "^0.17.6"
pre-commit = "^2.19.0"
pytest = "^7.1.2"
numpy = "^1.22.4"
scikit-learn = "^1.1.1"
pandas = "^1.4.2"
tensorflow = "^2.9.1"
tensorflow-datasets = "^4.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
